# wasm: use llvm-12 toolchain, bump wabt/binaryen