There's more changes to the wasm files than in the last bump, 11.0.0 -> 11.1.0; so, it's not entirely straightforward to say what exactly has changed.

That said, our linking and optimization machinery is still intact, and the benchmarks we have don't indicate anything catastrophic:

```
name                                    old time/op  new time/op  delta
WASMArrayIteration/10-16                 423µs ± 5%   405µs ± 4%    ~     (p=0.095 n=5+5)
WASMArrayIteration/100-16                441µs ± 8%   414µs ± 4%  -5.98%  (p=0.032 n=5+5)
WASMArrayIteration/1000-16               572µs ± 2%   559µs ± 2%    ~     (p=0.056 n=5+5)
WASMArrayIteration/10000-16             2.18ms ± 1%  2.35ms ± 6%  +8.11%  (p=0.008 n=5+5)
WASMSetIteration/10-16                   443µs ± 3%   421µs ± 3%  -4.98%  (p=0.008 n=5+5)
WASMSetIteration/100-16                  452µs ± 4%   427µs ± 3%  -5.48%  (p=0.008 n=5+5)
WASMSetIteration/1000-16                 672µs ± 2%   644µs ± 2%  -4.24%  (p=0.008 n=5+5)
WASMSetIteration/10000-16               3.66ms ± 5%  3.46ms ± 5%    ~     (p=0.056 n=5+5)
WASMObjectIteration/10-16                440µs ± 6%   400µs ± 4%  -9.18%  (p=0.008 n=5+5)
WASMObjectIteration/100-16               463µs ± 5%   433µs ± 7%    ~     (p=0.056 n=5+5)
WASMObjectIteration/1000-16              716µs ± 3%   722µs ± 2%    ~     (p=0.421 n=5+5)
WASMObjectIteration/10000-16            3.93ms ±11%  4.14ms ± 3%    ~     (p=0.151 n=5+5)
WASMLargeJSON/10x10-16                   483µs ± 4%   470µs ± 1%    ~     (p=0.421 n=5+5)
WASMLargeJSON/10x100-16                  491µs ± 5%   491µs ± 2%    ~     (p=1.000 n=5+5)
WASMLargeJSON/10x1000-16                 880µs ± 2%   858µs ± 4%    ~     (p=0.222 n=5+5)
WASMLargeJSON/10x10000-16               5.56ms ± 7%  5.65ms ± 7%    ~     (p=0.548 n=5+5)
WASMLargeJSON/100x100-16                 686µs ± 3%   711µs ± 4%  +3.62%  (p=0.032 n=5+5)
WASMLargeJSON/100x1000-16               3.17ms ± 5%  3.13ms ± 6%    ~     (p=0.421 n=5+5)
WASMVirtualDocs/total=1/hit=1-16         426µs ± 6%   415µs ± 2%    ~     (p=0.151 n=5+5)
WASMVirtualDocs/total=10/hit=1-16        416µs ± 3%   415µs ± 2%    ~     (p=1.000 n=5+5)
WASMVirtualDocs/total=100/hit=1-16       437µs ± 5%   421µs ± 6%    ~     (p=0.151 n=5+5)
WASMVirtualDocs/total=1000/hit=1-16      492µs ± 5%   460µs ± 2%  -6.60%  (p=0.008 n=5+5)
WASMVirtualDocs/total=10/hit=10-16       424µs ± 3%   421µs ± 3%    ~     (p=0.841 n=5+5)
WASMVirtualDocs/total=100/hit=10-16      428µs ± 3%   418µs ± 4%    ~     (p=0.095 n=5+5)
WASMVirtualDocs/total=1000/hit=10-16     487µs ± 5%   474µs ± 4%    ~     (p=0.222 n=5+5)
WASMVirtualDocs/total=100/hit=100-16     465µs ± 3%   442µs ± 5%  -4.95%  (p=0.032 n=5+5)
WASMVirtualDocs/total=1000/hit=100-16    523µs ± 5%   496µs ± 4%    ~     (p=0.056 n=5+5)
WASMVirtualDocs/total=1000/hit=1000-16   804µs ± 1%   826µs ± 1%  +2.64%  (p=0.008 n=5+5)
```