@srenatus 2021-05-06
This won't pass because the tests are building the image on the fly and timeout. After merging, the image will be built and pushed to dockerhub, and everything will be as good as new. So, I'll go ahead and get this in.